% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfiles_function.r
\name{loadfiles}
\alias{loadfiles}
\title{generate folder function}
\usage{
loadfiles(count_file, target_file)
}
\arguments{
\item{count_file}{count matrix raw or normalized}

\item{target_file}{file with experimental information}
}
\description{
This loads count and target file info (internal function only)
}
\keyword{loadsfiles}
