% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_normalize_raw_counts_function.r
\name{s1_normalize_raw_counts}
\alias{s1_normalize_raw_counts}
\title{s1 normalize raw counts function following Gale lab protocol}
\usage{
s1_normalize_raw_counts(countfile, targetfile, target_class = 10,
  blocking_column = FALSE, visualize_data = TRUE,
  filter_genes_below_counts = 0, figres = 100)
}
\arguments{
\item{countfile}{raw counts table (generally output by htseq).}

\item{targetfile}{target file.}

\item{target_class}{columns from the target file to build design matrix for future DE analysis}

\item{filter_genes_below_counts}{filter out genes with counts below a certain value, (default set to 0).}

\item{figres}{resolution at which to output figures (default is 300).}

\item{vizualize_data}{whether or not to generate figures (default set to true).}
}
\description{
This function allows the normalization of raw counts following a general protocol developed by previous members of the gale lab
}
\examples{
normalize_raw_counts(count_file.txt, target_file.csv, vizualize_data=TRUE, FilterGenesWithCounts=100)
}
\keyword{expression}
\keyword{gene}
\keyword{normalization}
