% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_feature_reduction_function.r
\name{s2_feature_reduction}
\alias{s2_feature_reduction}
\title{s2 feature reduction}
\usage{
s2_feature_reduction(countfile, targetfile, figres = 100)
}
\arguments{
\item{countfile}{raw counts table (generally output by htseq).}

\item{targetfile}{target file.}

\item{figres}{resolution at which to output figures (default is 300).}
}
\description{
This function does feature reduction and visualization on normalized voom counts
}
\examples{
s2_feature_reduction(count_file.txt, target_file.csv, vizualize_data=TRUE, FilterGenesWithCounts=100)
}
\keyword{expression}
\keyword{feature}
\keyword{gene}
\keyword{reduction}
