% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_DE_analysis_function.r
\name{s3_DE_analysis}
\alias{s3_DE_analysis}
\title{s3 differential gene analysis}
\usage{
s3_DE_analysis(
  targetfile = "./p1_modified_count_matrix_results/target_file.csv",
  folder2voomobjects = "s1_norm_raw_counts_results",
  results_folder = FALSE,
  gene_conversion_file = FALSE,
  blocking_column = FALSE,
  matrixfile = FALSE,
  pvalue = 0.05,
  logfoldchange = 1.5,
  reset_design_matrix = FALSE,
  target_columns = FALSE,
  batch_column = FALSE
)
}
\arguments{
\item{targetfile}{target file.}

\item{folder2voomobjects}{folder where necessary voom R objects are stored (should be results folder from step 1) defualt is s1_norm_raw_counts_results}

\item{results_folder}{User can specify name of output folder they want results to go in (default is s3_DE_results/)}

\item{gene_conversion_file}{file with alternative gene names (we usually convert Rhesus Ensembl genes to HGNC)}

\item{blocking_column}{column to account sampling from the same animal multiple times (needs to be the same as whatever was specified in step s1_normalize_raw_counts)}

\item{matrixfile}{text file outlining different contrasts to do in DE analysis}

\item{pvalue}{parameter for determining significantly DE genes (default 0.05)}

\item{logfoldchange}{parameter for determining significantly DE genes (default 1.5)}

\item{reset_design_matrix}{parameter for user to reset design matrix if they want to do this differently than they did in step 1 (defualt is FALSE)}

\item{target_columns}{vector of columns from target file to use as treatment group, this needs to be set if reset_design_matrix is TRUE}

\item{batch_column}{column from target file to be used as batch effect if reseting design matrix}
}
\description{
This function runs differential gene analysis (linear modeling on normalized counts)
}
\examples{
s3_DE_analysis(countfile="./s1_norm_raw_counts_results/1.norm_matrix.txt", targetfile="./p1_modified_count_matrix_results/target_file.csv", gene_conversion_file="rhesus2human.csv", blocking_column=2, matrixfile="./MATRIX.txt")
}
\keyword{differential}
\keyword{expression}
\keyword{linear}
\keyword{modeling}
