% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_feature_reduction_function.r
\name{s2_feature_reduction}
\alias{s2_feature_reduction}
\title{s2 feature reduction}
\usage{
s2_feature_reduction(countfile, targetfile, target_class = c(2, 5),
  figres = 100, pcva = TRUE, pca = TRUE, UMAP = TRUE,
  base_file_name = "vnorm.png")
}
\arguments{
\item{countfile}{normalized counts table (generally should have been generated by first step s1_normalize_raw_counts, but also this function can be run on raw cout file).}

\item{targetfile}{target file.}

\item{target_class}{columns from the target file to label samples on the pca/umap plots (has to be 2)}

\item{figres}{resolution at which to output figures (default is 300).}

\item{pcva}{run principal component variance analysis (default set to TRUE)}

\item{pca}{run principal component analysis (and singular vector decomposition) (default set to TRUE)}

\item{UMAP}{run Uniform Manifold Aproximation Projection (default set to TRUE)}

\item{base_file_name}{file name for all figures}
}
\description{
This function does feature reduction and visualization on normalized voom counts
}
\examples{
s2_feature_reduction(count_file.txt, target_file.csv, vizualize_data=TRUE, FilterGenesWithCounts=100)
}
\keyword{expression}
\keyword{feature}
\keyword{gene}
\keyword{reduction}
