% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_DE_analysis.r
\name{s3_DE_analysis}
\alias{s3_DE_analysis}
\title{s3 differential gene analysis}
\usage{
s3_DE_analysis()
}
\description{
This function runs differential gene analysis (linear modeling on normalized counts)
}
\examples{

}
\keyword{analysis}
\keyword{differential}
\keyword{gene}
