% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_DE_analysis_function.r
\name{s3_DE_analysis}
\alias{s3_DE_analysis}
\title{s3 differential gene analysis}
\usage{
s3_DE_analysis(
  countfile = "./s1_norm_raw_counts_results/1.norm_matrix.txt",
  targetfile = "./p1_modified_count_matrix_results/target_file.csv",
  gene_conversion_file = FALSE,
  blocking_column = FALSE,
  matrixfile = FALSE,
  pvalue = 0.05,
  logfoldchange = 1.5
)
}
\arguments{
\item{countfile}{normalized counts table (generally should have been generated by first step s1_normalize_raw_counts, but also this function can be run on raw cout file).}

\item{targetfile}{target file.}

\item{gene_conversion_file}{file with alternative gene names (we usually convert Rhesus Ensembl genes to HGNC)}

\item{blocking_column}{column to account sampling from the same animal multiple times (needs to be the same as whatever was specified in step s1_normalize_raw_counts)}

\item{matrixfile}{text file outlining different contrasts to do in DE analysis}

\item{pvalue}{parameter for determining significantly DE genes (default 0.05)}

\item{logfoldchange}{parameter for determining significantly DE genes (default 1.5)}
}
\description{
This function runs differential gene analysis (linear modeling on normalized counts)
}
\examples{
s3_DE_analysis(countfile='./s1_norm_raw_counts_results/1.norm_matrix.txt', targetfile='./p1_modified_count_matrix_results/target_file.csv', gene_conversion_file='rhesus2human.csv', blocking_column=2, matrixfile='./MATRIX.txt')
}
\keyword{differential}
\keyword{expression}
\keyword{linear}
\keyword{modeling}
