% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_gene_enrichment_analysis_function.r
\name{s4_gene_enrichment_analysis}
\alias{s4_gene_enrichment_analysis}
\title{s4 gene enrichment analysis}
\usage{
s4_gene_enrichment_analysis(
  DEgenes = "./s3_DE_results/3.ExpressMatrix_separate_LFC_HGNC_AV.csv",
  go_enrich_type = "BP",
  result_folder = FALSE,
  gene_name_column = 3,
  log_values_column = 7,
  pvalue = 0.1,
  NumTopGoTerms = 10,
  figres = 300,
  base_file_name = "ge.png"
)
}
\arguments{
\item{DEgenes}{file containg log fold change for DE genes (default ./s3_DE_results/3.ExpressMatrix_separate_LFC_HGNC_AV.csv)}

\item{go_enrich_type}{type of GO enrichment to do (BP, CC, or MF) (default BP)}

\item{result_folder}{user specified output folder (default is s4_gene_enrichment_results)}

\item{gene_name_column}{column in DEgenes file with HGNC symbols (default 3)}

\item{log_values_column}{column with log fold change values for GSEA (default 7)}

\item{pvalue}{adjusted pvalue cutoff (default 0.1)}

\item{NumTopGoTerms}{top GO terms to show (default 10)}

\item{figres}{resolution of output figures (default 300)}

\item{base_file_name}{name to save files under (default ge.png)}
}
\description{
This function runs gene enrichment analysis
}
\examples{
s4_gene_enrichment_analysis(DEgenes='./s3_DE_results/3.ExpressMatrix_separate_LFC_HGNC_AV.csv', go_enrich_type='BP', gene_name_column=3, log_values_column=4, pvalue=0.05, NumbTopGoTerms=10)
}
\keyword{analysis}
\keyword{enrichment}
\keyword{gene}
