% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_gene_enrichment_analysis_function.r
\name{s4_gene_enrichment_analysis}
\alias{s4_gene_enrichment_analysis}
\title{s4 gene enrichment analysis}
\usage{
s4_gene_enrichment_analysis(
  DEgenes = "./s3_DE_results/3.ExpressMatrix_separate_LFC_HGNC_AV.csv",
  log_values_column = 7,
  pvalue = 0.1,
  NumTopGoTerms = 10,
  figres = 100
)
}
\description{
This function runs gene enrichment analysis
}
\examples{

}
\keyword{analysis}
\keyword{enrichment}
\keyword{gene}
