% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_gene_enrichment_analysis_function.r
\name{s4_gene_enrichment_analysis}
\alias{s4_gene_enrichment_analysis}
\title{s4 gene enrichment analysis}
\usage{
s4_gene_enrichment_analysis(
  go_enrich_type = "BP",
  universe = TRUE,
  universe_file = " ./s1_raw_counts_results/1.norm_matrix_HGNC.txt",
  DEgenes = FALSE,
  log_values_column = FALSE,
  rnkfile = FALSE,
  results_folder = "s4_gene_enrichment_results",
  comparison = FALSE,
  modules = TRUE,
  module_results_folder = "s3_DE_results",
  NumTopGoTerms = 30,
  figres = 300,
  ensembl_retrieve = TRUE,
  gene_name_type = "SYMBOL",
  base_file_name = "ge.png"
)
}
\arguments{
\item{go_enrich_type}{type of GO enrichment to do (BP, CC, or MF) (default BP)}

\item{universe}{Will specify whether you want to use the whole genome as background (FALSE) or set of total expressed genes(TRUE)}

\item{universe_file}{specifies path to universe file(defualt ./s1_raw_counts_results/1.norm_matrix_HGNC.txt)}

\item{DEgenes}{file containg log fold change for DE genes (default FALSE, needs to be given with log_values_column if used)}

\item{rnkfile}{file containing list of users own genes (not generated in step s3).  Must be in rank file format (gene name & logfold value seperated by tab)}

\item{results_folder}{user specified output folder (default is s4_gene_enrichment_results)}

\item{comparison}{specify name of time point comparison to perform over enrichment analysis}

\item{modules}{perform over enrichment analysis on modules generated in step s3}

\item{module_results_folder}{folder where DE results are (default s3_DE_results/)}

\item{NumTopGoTerms}{top GO terms to show (default 10)}

\item{figres}{resolution of output figures (default 300)}

\item{ensembl_retrieve}{whether or not to retrieve gene name descriptions}

\item{base_file_name}{name to save files under (default ge.png)}

\item{log_fold_column}{column to pull from DEgenes}
}
\description{
This function runs gene enrichment analysis
}
\examples{
s4_gene_enrichment_analysis(DEgenes=FALSE, go_enrich_type="BP", log_values_column=FALSE, modules=TRUE, pvalue=0.05, qvalue=0.05, NumTopGoTerms=30)
}
\keyword{analysis}
\keyword{enrichment}
\keyword{gene}
